refactor(core): split inventory into items, AC, and equip modules
Keep core.inventory as a thin facade for stable imports.
